Estudo/dashboard/requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
